feat: rename internal/doctorv2 to internal/doctor; strip V2 vocabulary — task 2 of 8 (#538)
Renames the doctorv2 package to doctor and updates callers in
internal/cli/check.go and internal/cli/repair.go. Package doc-strings
now describe the package as 'the doctor health check' rather than 'the
v2 doctor health check', and the stale 'checks the v2 framework mount
state' comment is shortened.
